Split WalletService
