+++ src/app/components/pages/send-skycoin/send-verify/send-verify.component.spec.ts
@@ -3,8 +3,9 @@ import { MatSnackBar } from '@angular/material';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
 
 import { SendVerifyComponent } from './send-verify.component';
-import { WalletService } from '../../../../services/wallet.service';
-import { MockTranslatePipe, MockWalletService, MockMatSnackBar } from '../../../../utils/test-mocks';
+import { BalanceService } from '../../../../services/wallet/balance.service';
+import { SpendingService } from '../../../../services/wallet/spending.service';
+import { MockTranslatePipe, MockBalanceService, MockSpendingService, MockMatSnackBar } from '../../../../utils/test-mocks';
 
 describe('SendVerifyComponent', () => {
   let component: SendVerifyComponent;
@@ -15,7 +16,8 @@ describe('SendVerifyComponent', () => {
       declarations: [ SendVerifyComponent, MockTranslatePipe ],
       schemas: [ NO_ERRORS_SCHEMA ],
       providers: [
-        { provide: WalletService, useClass: MockWalletService },
+        { provide: BalanceService, useClass: MockBalanceService },
+        { provide: SpendingService, useClass: MockSpendingService },
         { provide: MatSnackBar, useClass: MockMatSnackBar }
       ]
     })
