+++ src/app/components/pages/wallets/wallet-detail/wallet-detail.component.ts
@@ -4,7 +4,7 @@ import { MatSnackBar, MatSnackBarConfig } from '@angular/material';
 import { TranslateService } from '@ngx-translate/core';
 
 import { ConfirmationData, Wallet, Address } from '../../../../app.datatypes';
-import { WalletService } from '../../../../services/wallet.service';
+import { WalletService } from '../../../../services/wallet/wallet.service';
 import { QrCodeComponent } from '../../../layout/qr-code/qr-code.component';
 import { ChangeNameComponent } from '../change-name/change-name.component';
 import { openUnlockWalletModal } from '../../../../utils/index';
